move dir, test=doc
