+++ paddlespeech/server/bin/main.py
@@ -16,10 +16,10 @@ import uvicorn
 import yaml
 from fastapi import FastAPI
 
-from engine.engine_factory import EngineFactory
-from restful.api import setup_router
-from utils.config import get_config
-from utils.log import logger
+from paddlespeech.server.engine.engine_factory import EngineFactory
+from paddlespeech.server.restful.api import setup_router
+from paddlespeech.server.utils.config import get_config
+from paddlespeech.server.utils.log import logger
 
 app = FastAPI(
     title="PaddleSpeech Serving API", description="Api", version="0.0.1")
