+++ paddlespeech/server/engine/engine_factory.py
@@ -11,3 +11,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from typing import Text
+
+from paddlespeech.server.engine.asr.python.asr_engine import ASREngine
+#from paddlespeech.server.engine.tts.python.tts_engine import TTSEngine
+from paddlespeech.server.engine.tts.paddleinference.tts_engine import TTSEngine
+
+
+__all__ = ['EngineFactory']
+
+
+class EngineFactory(object):
+    @staticmethod
+    def get_engine(engine_name: Text):
+        if engine_name == 'asr':
+            return ASREngine()
+        elif engine_name == 'tts':
+            return TTSEngine()
+        else:
+            return None
